ADD interface and implementation to mongobackend
